Prevent duplicate empty votes in edge case scenario
When the vote is cast exactly at the same time on two separate
workers/servers, there can be an extra "empty" vote recorded for
the user. The last vote is always the "truth" as the previous vote
will not have any orders after they are mapped with the latest
vote.