Implement text search through linked plans
Allow the free text search to search through the projects
themselves as well as the linked plans in order to make it easier
for participants to find the project that has been created based
on the plan they initially submitted. For example, the plan's
title might have changed to something else when it might not be
easy to find when searching with the original title.